+++ construct_graph.h
@@ -4,10 +4,10 @@
 #define EPS 2
 
 void construct_fully_connected_matrix(double *points, int lines, int dim, double *ret);
-void construct_eps_neighborhood_matrix(double *points, int lines, int dim, int *ret);
+void construct_eps_neighborhood_matrix(double *points, int lines, int dim, double *ret);
 void construct_normalized_laplacian_sym_matrix(double *weighted_adj_matrix, int num_points, double *ret);
 void construct_normalized_laplacian_rw_matrix(double *weighted_adj_matrix, int num_points, double *ret);
 void construct_unnormalized_laplacian(double *graph, int n, double *ret);
-void construct_knn_matrix(double *points, int lines, int dim, int k, int *ret);
+void construct_knn_matrix(double *points, int lines, int dim, int k, double *ret);
 
 #endif
